Separated test class into another test class.
