+++ Tests/AttributesTests/validators/ValidationPushProtocolTests.swift
@@ -57,8 +57,6 @@
 @testable import Attributes
 import XCTest
 
-// swiftlint:disable type_body_length
-
 /// Test class for ``ValidationPushProtocol`` default implementations.
 final class ValidationPushProtocolTests: XCTestCase {
 
@@ -267,233 +265,4 @@ final class ValidationPushProtocolTests: XCTestCase {
         }
     }
 
-    /// Test equals function.
-    func testEquals() throws {
-        let readPath = ReadOnlyPath(EquatablePoint.self)
-        let point = EquatablePoint(x: 1, y: 2)
-        let validator = NullValidator<EquatablePoint>()
-        let path = TestValidationPath(path: readPath).push {root, _ in
-            try validator.performValidation(root)
-        }
-        .equals(point)
-        try path.performValidation(point)
-        XCTAssertEqual(validator.timesCalled, 1)
-        XCTAssertEqual(validator.lastParameter, point)
-        let point2 = EquatablePoint(x: 2, y: 3)
-        XCTAssertThrowsError(try path.performValidation(point2)) {
-            guard let error = $0 as? AttributeError<EquatablePoint> else {
-                XCTFail("Incorrect error thrown.")
-                return
-            }
-            XCTAssertEqual(error.message, "Must equal \(point).")
-            XCTAssertEqual(error.path, AnyPath(path.path))
-        }
-        XCTAssertEqual(validator.timesCalled, 2)
-        XCTAssertEqual(validator.lastParameter, point2)
-    }
-
-    /// Test notEquals function.
-    func testNotEquals() throws {
-        let readPath = ReadOnlyPath(EquatablePoint.self)
-        let point = EquatablePoint(x: 1, y: 2)
-        let point2 = EquatablePoint(x: 2, y: 3)
-        let validator = NullValidator<EquatablePoint>()
-        let path = TestValidationPath(path: readPath).push {root, _ in
-            try validator.performValidation(root)
-        }
-        .notEquals(point)
-        try path.performValidation(point2)
-        XCTAssertEqual(validator.timesCalled, 1)
-        XCTAssertEqual(validator.lastParameter, point2)
-        XCTAssertThrowsError(try path.performValidation(point)) {
-            guard let error = $0 as? AttributeError<EquatablePoint> else {
-                XCTFail("Incorrect error thrown.")
-                return
-            }
-            XCTAssertEqual(error.message, "Must not equal \(point).")
-            XCTAssertEqual(error.path, AnyPath(path.path))
-        }
-        XCTAssertEqual(validator.timesCalled, 2)
-        XCTAssertEqual(validator.lastParameter, point)
-    }
-
-    /// Test equalsTrue method.
-    func testEqualsTrue() throws {
-        let readPath = ReadOnlyPath(Bool.self)
-        let value = true
-        let notValue = false
-        let validator = NullValidator<Bool>()
-        let path = TestValidationPath(path: readPath).push {root, _ in
-            try validator.performValidation(root)
-        }
-        .equalsTrue()
-        try path.performValidation(value)
-        XCTAssertEqual(validator.timesCalled, 1)
-        XCTAssertEqual(validator.lastParameter, value)
-        XCTAssertThrowsError(try path.performValidation(notValue)) {
-            guard let error = $0 as? AttributeError<Bool> else {
-                XCTFail("Incorrect error thrown.")
-                return
-            }
-            XCTAssertEqual(error.message, "Must equal \(value).")
-            XCTAssertEqual(error.path, AnyPath(path.path))
-        }
-        XCTAssertEqual(validator.timesCalled, 2)
-        XCTAssertEqual(validator.lastParameter, notValue)
-    }
-
-    /// Test equalsFalse method.
-    func testEqualsFalse() throws {
-        let readPath = ReadOnlyPath(Bool.self)
-        let value = true
-        let notValue = false
-        let validator = NullValidator<Bool>()
-        let path = TestValidationPath(path: readPath).push {root, _ in
-            try validator.performValidation(root)
-        }
-        .equalsFalse()
-        try path.performValidation(notValue)
-        XCTAssertEqual(validator.timesCalled, 1)
-        XCTAssertEqual(validator.lastParameter, notValue)
-        XCTAssertThrowsError(try path.performValidation(value)) {
-            guard let error = $0 as? AttributeError<Bool> else {
-                XCTFail("Incorrect error thrown.")
-                return
-            }
-            XCTAssertEqual(error.message, "Must equal \(notValue).")
-            XCTAssertEqual(error.path, AnyPath(path.path))
-        }
-        XCTAssertEqual(validator.timesCalled, 2)
-        XCTAssertEqual(validator.lastParameter, value)
-    }
-
-    /// Test between method.
-    func testBetween() throws {
-        let readPath = ReadOnlyPath(Int.self)
-        let validator = NullValidator<Int>()
-        let path = TestValidationPath(path: readPath).push { root, _ in
-            try validator.performValidation(root)
-        }
-        .between(min: 3, max: 7)
-        try path.performValidation(5)
-        XCTAssertEqual(validator.timesCalled, 1)
-        XCTAssertEqual(validator.lastParameter, 5)
-        try path.performValidation(3)
-        XCTAssertEqual(validator.timesCalled, 2)
-        XCTAssertEqual(validator.lastParameter, 3)
-        try path.performValidation(7)
-        XCTAssertEqual(validator.timesCalled, 3)
-        XCTAssertEqual(validator.lastParameter, 7)
-        XCTAssertThrowsError(try path.performValidation(1)) {
-            guard let error = $0 as? ValidationError<Int> else {
-                XCTFail("Incorrect error thrown.")
-                return
-            }
-            XCTAssertEqual(error.message, "Must be between \(3) and \(7).")
-            XCTAssertEqual(error.path, AnyPath(path.path))
-        }
-        XCTAssertEqual(validator.timesCalled, 4)
-        XCTAssertEqual(validator.lastParameter, 1)
-    }
-
-    /// Test `lessThan` method.
-    func testLessThan() throws {
-        let readPath = ReadOnlyPath(Int.self)
-        let validator = NullValidator<Int>()
-        let path = TestValidationPath(path: readPath).push { root, _ in
-            try validator.performValidation(root)
-        }
-        .lessThan(5)
-        try path.performValidation(4)
-        XCTAssertEqual(validator.timesCalled, 1)
-        XCTAssertEqual(validator.lastParameter, 4)
-        XCTAssertThrowsError(try path.performValidation(5)) {
-            guard let error = $0 as? ValidationError<Int> else {
-                XCTFail("Incorrect error thrown.")
-                return
-            }
-            XCTAssertEqual(error.message, "Must be less than \(5).")
-            XCTAssertEqual(error.path, AnyPath(path.path))
-        }
-        XCTAssertEqual(validator.timesCalled, 2)
-        XCTAssertEqual(validator.lastParameter, 5)
-    }
-
-    /// Test `lessThanEqual` method.
-    func testLessThanEqual() throws {
-        let readPath = ReadOnlyPath(Int.self)
-        let validator = NullValidator<Int>()
-        let path = TestValidationPath(path: readPath).push { root, _ in
-            try validator.performValidation(root)
-        }
-        .lessThanEqual(5)
-        try path.performValidation(4)
-        XCTAssertEqual(validator.timesCalled, 1)
-        XCTAssertEqual(validator.lastParameter, 4)
-        try path.performValidation(5)
-        XCTAssertEqual(validator.timesCalled, 2)
-        XCTAssertEqual(validator.lastParameter, 5)
-        XCTAssertThrowsError(try path.performValidation(6)) {
-            guard let error = $0 as? ValidationError<Int> else {
-                XCTFail("Incorrect error thrown.")
-                return
-            }
-            XCTAssertEqual(error.message, "Must be less than or equal to \(5).")
-            XCTAssertEqual(error.path, AnyPath(path.path))
-        }
-        XCTAssertEqual(validator.timesCalled, 3)
-        XCTAssertEqual(validator.lastParameter, 6)
-    }
-
-    /// Test `greaterThan` method.
-    func testGreaterThan() throws {
-        let readPath = ReadOnlyPath(Int.self)
-        let validator = NullValidator<Int>()
-        let path = TestValidationPath(path: readPath).push { root, _ in
-            try validator.performValidation(root)
-        }
-        .greaterThan(5)
-        try path.performValidation(6)
-        XCTAssertEqual(validator.timesCalled, 1)
-        XCTAssertEqual(validator.lastParameter, 6)
-        XCTAssertThrowsError(try path.performValidation(5)) {
-            guard let error = $0 as? ValidationError<Int> else {
-                XCTFail("Incorrect error thrown.")
-                return
-            }
-            XCTAssertEqual(error.message, "Must be greater than \(5).")
-            XCTAssertEqual(error.path, AnyPath(path.path))
-        }
-        XCTAssertEqual(validator.timesCalled, 2)
-        XCTAssertEqual(validator.lastParameter, 5)
-    }
-
-    /// Test `greaterThanEqual` method.
-    func testGreaterThanEqual() throws {
-        let readPath = ReadOnlyPath(Int.self)
-        let validator = NullValidator<Int>()
-        let path = TestValidationPath(path: readPath).push { root, _ in
-            try validator.performValidation(root)
-        }
-        .greaterThanEqual(5)
-        try path.performValidation(5)
-        XCTAssertEqual(validator.timesCalled, 1)
-        XCTAssertEqual(validator.lastParameter, 5)
-        try path.performValidation(6)
-        XCTAssertEqual(validator.timesCalled, 2)
-        XCTAssertEqual(validator.lastParameter, 6)
-        XCTAssertThrowsError(try path.performValidation(4)) {
-            guard let error = $0 as? ValidationError<Int> else {
-                XCTFail("Incorrect error thrown.")
-                return
-            }
-            XCTAssertEqual(error.message, "Must be greater than or equal to \(5).")
-            XCTAssertEqual(error.path, AnyPath(path.path))
-        }
-        XCTAssertEqual(validator.timesCalled, 3)
-        XCTAssertEqual(validator.lastParameter, 4)
-    }
-
 }
-
-// swiftlint:enable type_body_length
